Re-write of the existing manifest in yaml.
Some minor tweaks done on the way such as simplified archive handling and moving the executable file away from the flatpak lib folder.

The TMPDIR finish arg is also removed, as I didnt see it used anywhere.

A small Makefile is added for simpler usage of flatpak and flatpak-builder when developing.

Finally, the gitignore is trimmed to exclude temporary flatpak build files.