Unpack archive directly to the destination location
As a consequence of this, we also install the icons from the installation location. This is done in a loop as to save some lines and avoid repetition. Another benefit of a loop is if at some point later 256 or 128 icons are added, this is very easy to add.

Smaller manifest tweaks 3/n